+++ app/controllers/controller_mixin.rb
@@ -1,6 +1,6 @@
 module ControllerMixin
   def self.included(base)
-    base.helper_method :current_user
+    base.helper_method :"current_#{QuoVadis.model_instance_name}"
   end
 
   protected
@@ -12,50 +12,60 @@ module ControllerMixin
 
   private
 
-  # Remembers the authenticated <tt>user</tt> (in this session and future sessions).
-  #
-  # If you want to sign in a <tt>user</tt> you have just created, call <tt>sign_in</tt>
-  # instead.
-  def current_user=(user)
-    remember_user_in_session user
-    remember_user_between_sessions user
-  end
-
-  # Returns the authenticated user.
-  def current_user
-    @current_user ||= find_authenticated_user
-  end
-
-  # Does nothing if we already have an authenticated user.  If we don't have an
-  # authenticated user, it stores the desired URL and redirects to the sign in URL.
-  def authenticate
-    unless current_user
-      session[:quo_vadis_original_url] = request.fullpath
-      flash[:notice] = t('quo_vadis.flash.sign_in.before') unless t('quo_vadis.flash.sign_in.before').blank?
-      redirect_to sign_in_url
+  class_eval <<-END, __FILE__, __LINE__ + 1
+    # Remembers the authenticated <tt>user</tt> (in this session and future sessions).
+    #
+    # If you want to sign in a <tt>user</tt> you have just created, call <tt>sign_in</tt>
+    # instead.
+    def current_#{QuoVadis.model_instance_name}=(user)
+      remember_user_in_session user
+      remember_user_between_sessions user
     end
-  end
 
-  # Signs in a user, i.e. remembers them in the session, runs the sign-in hook,
-  # and redirects appropriately.
-  #
-  # This method should be called when you have just authenticated a <tt>user</tt>
-  # and you need to sign them in.  For example, if a new user has just signed up,
-  # you should call this method to sign them in.
-  def sign_in(user)
-    prevent_session_fixation
-    self.current_user = user
-    QuoVadis.signed_in_hook user, self
-    redirect_to QuoVadis.signed_in_url(user, original_url)
-  end
+    # Returns the authenticated user.
+    def current_#{QuoVadis.model_instance_name}
+      @current_#{QuoVadis.model_instance_name} ||= find_authenticated_user
+    end
+
+    # Does nothing if we already have an authenticated user.  If we don't have an
+    # authenticated user, it stores the desired URL and redirects to the sign in URL.
+    def authenticate
+      unless current_#{QuoVadis.model_instance_name}
+        session[:quo_vadis_original_url] = request.fullpath
+        flash[:notice] = t('quo_vadis.flash.sign_in.before') unless t('quo_vadis.flash.sign_in.before').blank?
+        redirect_to sign_in_url
+      end
+    end
+
+    # Signs in a user, i.e. remembers them in the session, runs the sign-in hook,
+    # and redirects appropriately.
+    #
+    # This method should be called when you have just authenticated a <tt>user</tt>
+    # and you need to sign them in.  For example, if a new user has just signed up,
+    # you should call this method to sign them in.
+    def sign_in(user)
+      prevent_session_fixation
+      self.current_#{QuoVadis.model_instance_name} = user
+      QuoVadis.signed_in_hook user, self
+      redirect_to QuoVadis.signed_in_url(user, original_url)
+    end
+
+    def remember_user_in_session(user) # :nodoc:
+      session[:current_#{QuoVadis.model_instance_name}_id] = user ? user.id : nil
+    end
+
+    def find_user_by_cookie # :nodoc:
+      #{QuoVadis.model}.find_by_salt(*cookies.signed[:remember_me]) if cookies.signed[:remember_me]
+    end
+
+    def find_user_by_session # :nodoc:
+      #{QuoVadis.model}.find(session[:current_#{QuoVadis.model_instance_name}_id]) if session[:current_#{QuoVadis.model_instance_name}_id]
+    end
+  END
 
   # Returns true if the sign-in process is blocked to the user, false otherwise.
   def blocked?
     QuoVadis.blocked?(self)
-  end
-
-  def remember_user_in_session(user) # :nodoc:
-    session[:current_user_id] = user ? user.id : nil
   end
 
   def remember_user_between_sessions(user) # :nodoc:
@@ -72,14 +82,6 @@ module ControllerMixin
 
   def find_authenticated_user # :nodoc:
     find_user_by_session || find_user_by_cookie
-  end
-
-  def find_user_by_cookie # :nodoc:
-    User.find_by_salt(*cookies.signed[:remember_me]) if cookies.signed[:remember_me]
-  end
-
-  def find_user_by_session # :nodoc:
-    User.find(session[:current_user_id]) if session[:current_user_id]
   end
 
   # Returns the URL if any which the user tried to visit before being forced to authenticate.